Add tests for calendar props/methods in date picker
